Refatorado(Material): dividido material.py por provider em módulos dedicados
- Criado material_common.py com helpers compartilhados (rotação de
  API key, redação de segredo em log, checagem de aspect ratio,
  registro de fonte do material)
- Criado material_pexels.py, material_pixabay.py, material_coverr.py
  e material_wavespeed.py, um por provider de busca/geração de vídeo
- Reduzido material.py de 1584 para 676 linhas, mantido como fachada
  e camada de orquestração de download/cache (download_videos,
  _search_videos_with_cache, save_video)
