+++ app/services/material.py
@@ -1,11 +1,8 @@
 import math
 import os
 import random
-import threading
-import time
-from pathlib import Path
+import time  # noqa: F401 (kept for `material.time` back-compat access, e.g. in tests)
 from typing import Any, Callable, List
-from urllib.parse import quote_plus, urlencode, urlsplit, urlunsplit
 
 import requests
 from loguru import logger
@@ -13,942 +10,37 @@ from moviepy.video.io.VideoFileClip import VideoFileClip
 
 from app.config import config
 from app.models.schema import MaterialInfo, VideoAspect, VideoConcatMode
-from app.services import material_cache, task_artifacts, volcengine_seedance
+from app.services import material_cache, task_artifacts, volcengine_seedance  # noqa: F401 (task_artifacts kept for back-compat access, e.g. in tests)
 from app.utils import utils
 
-# Thread-safe counter for API key rotation
-_api_key_counter = 0
-_api_key_lock = threading.Lock()
-
-
-def _safe_public_url(value: Any) -> str | None:
-    """
-    只保留可公开展示的 HTTP(S) 页面地址，并移除查询参数和凭据。
-
-    素材下载地址可能携带 API Key、签名 JWT 或临时 token。任务清单只需要
-    帮助用户回到供应商的公开素材页，不应保存鉴权参数；用户信息形式的 URL
-    同样拒绝，避免 ``https://user:pass@example.com`` 一类内容落盘。
-    """
-    if not isinstance(value, str) or not value.strip():
-        return None
-
-    try:
-        parsed = urlsplit(value.strip())
-    except ValueError:
-        return None
-    if (
-        parsed.scheme not in {"http", "https"}
-        or not parsed.hostname
-        or parsed.username is not None
-        or parsed.password is not None
-    ):
-        return None
-    return urlunsplit((parsed.scheme, parsed.netloc, parsed.path, "", ""))
-
-
-def _creator_info(value: Any) -> dict[str, str] | None:
-    """从不同供应商的作者结构中提取统一的公开字段。"""
-    if isinstance(value, str) and value.strip():
-        return {"name": value.strip()}
-    if not isinstance(value, dict):
-        return None
-
-    creator: dict[str, str] = {}
-    creator_id = value.get("id")
-    creator_name = value.get("name") or value.get("username")
-    creator_page = _safe_public_url(
-        value.get("url") or value.get("profile_url") or value.get("profile_page")
-    )
-    if creator_id is not None:
-        creator["id"] = str(creator_id)
-    if creator_name:
-        creator["name"] = str(creator_name)
-    if creator_page:
-        creator["profile_page"] = creator_page
-    return creator or None
-
-
-def _material_source_record(item: MaterialInfo, local_path: str) -> dict[str, Any]:
-    """
-    为成功下载的素材生成轻量来源记录。
-
-    ``source_info`` 可能来自缓存，甚至来自外部构造的 ``MaterialInfo``，因此
-    不能原样写入。这里按白名单重新构造，只保留公开页面、业务标识和尺寸，
-    并只记录本地文件名，避免用户目录或 Docker 挂载路径进入任务文件。
-    """
-    source = item.source_info if isinstance(item.source_info, dict) else {}
-    record: dict[str, Any] = {
-        "provider": str(item.provider or source.get("provider") or ""),
-        "local_file": Path(local_path).name,
-        "duration": int(item.duration),
-    }
-
-    search_term = source.get("search_term")
-    asset_id = source.get("asset_id")
-    source_page = _safe_public_url(source.get("source_page"))
-    if isinstance(search_term, str) and search_term.strip():
-        record["search_term"] = search_term.strip()
-    if asset_id not in (None, ""):
-        record["asset_id"] = str(asset_id)
-    if source_page:
-        record["source_page"] = source_page
-
-    creator = _creator_info(source.get("creator"))
-    if creator:
-        record["creator"] = creator
-
-    raw_rendition = source.get("rendition")
-    if isinstance(raw_rendition, dict):
-        rendition = {}
-        for field in ("id", "width", "height"):
-            value = raw_rendition.get(field)
-            if value not in (None, ""):
-                rendition[field] = str(value) if field == "id" else value
-        if rendition:
-            record["rendition"] = rendition
-    return record
-
-
-def _persist_material_sources(
-    task_id: str,
-    material_sources: list[dict[str, Any]],
-) -> None:
-    """
-    将当前实际下载成功的素材来源补充到任务清单。
-
-    任务记录是辅助能力，不能改变视频下载函数的返回值，也不能因为写盘失败
-    中断成片主流程。``patch_script_data`` 会负责原子替换和异常日志；这里仅在
-    成功后记录数量，便于确认任务追溯信息是否已经落盘。
-    """
-    try:
-        saved = task_artifacts.patch_script_data(
-            task_id,
-            material_sources=material_sources,
-        )
-        if saved:
-            logger.info(
-                f"saved material source records: "
-                f"task_id={task_id}, count={len(material_sources)}"
-            )
-    except Exception as exc:
-        # task_artifacts 自身已经按失败降级设计，这里仍保留最后一道隔离，
-        # 防止未来实现调整或目录解析异常意外影响素材下载返回值。
-        logger.warning(
-            "failed to persist material source records: "
-            f"task_id={task_id}, error={type(exc).__name__}, detail={exc}"
-        )
-
-
-def _get_tls_verify() -> bool:
-    # 默认开启 TLS 证书校验，防止素材搜索和下载过程被中间人篡改。
-    # 仅在企业代理、自签证书等明确需要的场景下，允许用户通过
-    # `config.toml` 显式设置 `tls_verify = false` 临时关闭。
-    tls_verify = config.app.get("tls_verify", True)
-    if isinstance(tls_verify, str):
-        tls_verify = tls_verify.strip().lower() not in ("0", "false", "no", "off")
-
-    if not tls_verify:
-        logger.warning(
-            "TLS certificate verification is disabled by config.app.tls_verify=false. "
-            "Only use this in trusted proxy environments."
-        )
-
-    return bool(tls_verify)
-
-
-def get_api_key(cfg_key: str):
-    api_keys = config.app.get(cfg_key)
-    if not api_keys:
-        raise ValueError(
-            f"\n\n##### {cfg_key} is not set #####\n\n"
-            f"Please set it in the config.toml file: {config.config_file}\n"
-        )
-
-    # if only one key is provided, return it
-    if isinstance(api_keys, str):
-        return api_keys
-
-    global _api_key_counter
-    with _api_key_lock:
-        _api_key_counter += 1
-        return api_keys[_api_key_counter % len(api_keys)]
-
-
-def _redact_secret(message: str, secret: str) -> str:
-    """
-    对即将写入日志的异常文本做最小范围脱敏。
-
-    requests 的连接异常可能包含完整请求 URL，而 Pixabay API Key 通过查询
-    参数传递。这里同时替换原始值和 URL 编码值，既保留网络错误信息用于排查，
-    又避免密钥进入日志文件。
-    """
-    safe_message = str(message)
-    if not secret:
-        return safe_message
-
-    safe_message = safe_message.replace(secret, "***")
-    encoded_secret = quote_plus(secret)
-    if encoded_secret != secret:
-        safe_message = safe_message.replace(encoded_secret, "***")
-    return safe_message
-
-
-def _redact_request_error(error: Exception, *secrets: str) -> str:
-    """
-    保留网络异常的可排查信息，同时移除 API Key 和代理凭据。
-
-    直接只记录异常类型会丢失 DNS、证书、超时等关键上下文；直接记录原始异常
-    又可能回显完整请求 URL。统一入口可以让三个素材供应商使用相同脱敏规则。
-    """
-    safe_message = str(error)
-    for secret in secrets:
-        safe_message = _redact_secret(safe_message, str(secret or ""))
-    for proxy_url in config.proxy.values():
-        safe_message = _redact_secret(safe_message, str(proxy_url))
-    return safe_message
-
-
-def _is_cloudflare_challenge(response: requests.Response) -> bool:
-    """
-    识别 Cloudflare 返回的 HTML Challenge，而不是把它当成 Pixabay JSON。
-
-    Cloudflare 通常会设置 `cf-mitigated: challenge`；部分部署只返回带有
-    "Just a moment" 或 challenge-platform 的 HTML，因此保留内容特征兜底。
-    响应正文仅在内存中判断，不写入日志，避免记录无价值的大段 HTML。
-    """
-    headers = getattr(response, "headers", {}) or {}
-    if str(headers.get("cf-mitigated", "")).lower() == "challenge":
-        return True
-
-    content_type = str(headers.get("content-type", "")).lower()
-    if "text/html" not in content_type:
-        return False
-
-    body = str(getattr(response, "text", "")).lower()
-    return "just a moment" in body or "/cdn-cgi/challenge-platform/" in body
-
-
-def _matches_video_aspect(
-    width: Any,
-    height: Any,
-    video_aspect: VideoAspect,
-    *,
-    is_vertical: Any = None,
-) -> bool:
-    """
-    判断远端素材是否与目标画面方向一致。
-
-    Pexels、Pixabay 和 Coverr 的响应字段并不统一，因此先使用宽高做可靠判断；
-    Coverr 部分历史响应缺少尺寸时，再使用明确的 ``is_vertical`` 布尔值兜底。
-    无法确认方向的素材直接跳过，避免竖屏任务混入横屏素材并在成片中产生黑边。
-    """
-    aspect = VideoAspect(video_aspect)
-    try:
-        normalized_width = int(float(width))
-        normalized_height = int(float(height))
-    except (TypeError, ValueError):
-        normalized_width = 0
-        normalized_height = 0
-
-    if normalized_width > 0 and normalized_height > 0:
-        if aspect == VideoAspect.portrait:
-            return normalized_height > normalized_width
-        if aspect == VideoAspect.landscape:
-            return normalized_width > normalized_height
-        return normalized_width == normalized_height
-
-    if isinstance(is_vertical, bool) and aspect != VideoAspect.square:
-        return is_vertical == (aspect == VideoAspect.portrait)
-    return False
-
-
-def _filter_materials_by_aspect(
-    items: List[MaterialInfo],
-    video_aspect: VideoAspect,
-) -> List[MaterialInfo]:
-    """
-    对缓存结果再次校验方向。
-
-    素材搜索缓存最长保留 24 小时，升级前写入的缓存可能包含方向不匹配的素材。
-    在统一缓存入口过滤可以让修复立即生效，也能防御第三方 Provider 或旧缓存
-    遗漏远端筛选。无法读取 rendition 尺寸的旧条目按未验证处理并跳过。
-    """
-    aspect = VideoAspect(video_aspect)
-    if aspect == VideoAspect.square:
-        # Pixabay 和 Coverr 很少提供原生方形素材。方形输出沿用既有行为，
-        # 接受可用候选并交给视频合成阶段裁剪，避免升级后 1:1 任务无素材。
-        return list(items)
-
-    filtered_items = []
-    for item in items:
-        source_info = item.source_info if isinstance(item.source_info, dict) else {}
-        rendition = source_info.get("rendition")
-        rendition = rendition if isinstance(rendition, dict) else {}
-        if _matches_video_aspect(
-            rendition.get("width"),
-            rendition.get("height"),
-            aspect,
-        ):
-            filtered_items.append(item)
-    return filtered_items
-
-
-def search_videos_pexels(
-    search_term: str,
-    minimum_duration: int,
-    video_aspect: VideoAspect = VideoAspect.portrait,
-) -> List[MaterialInfo]:
-    aspect = VideoAspect(video_aspect)
-    video_orientation = aspect.name
-    video_width, video_height = aspect.to_resolution()
-    api_key = get_api_key("pexels_api_keys")
-    headers = {
-        "Authorization": api_key,
-        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/115.0.0.0 Safari/537.36",
-    }
-    # Build URL
-    params = {"query": search_term, "per_page": 20, "orientation": video_orientation}
-    query_url = f"https://api.pexels.com/v1/videos/search?{urlencode(params)}"
-    logger.info(f"searching videos on pexels: term={search_term!r}")
-
-    try:
-        r = requests.get(
-            query_url,
-            headers=headers,
-            proxies=config.proxy,
-            verify=_get_tls_verify(),
-            timeout=(30, 60),
-        )
-        response = r.json()
-        video_items = []
-        if "videos" not in response:
-            logger.error("pexels video search returned an unsupported response")
-            return video_items
-        videos = response["videos"]
-        # loop through each video in the result
-        for v in videos:
-            duration = v["duration"]
-            # check if video has desired minimum duration
-            if duration < minimum_duration:
-                continue
-            video_files = v["video_files"]
-            # loop through each url to determine the best quality
-            for video in video_files:
-                w = int(video["width"])
-                h = int(video["height"])
-                if (
-                    _matches_video_aspect(w, h, aspect)
-                    and w == video_width
-                    and h == video_height
-                ):
-                    item = MaterialInfo()
-                    item.provider = "pexels"
-                    item.url = video["link"]
-                    item.duration = duration
-                    item.source_info = {
-                        "provider": "pexels",
-                        "search_term": search_term,
-                        "asset_id": (
-                            str(v.get("id")) if v.get("id") is not None else None
-                        ),
-                        "source_page": _safe_public_url(v.get("url")),
-                        "creator": _creator_info(v.get("user")),
-                        "rendition": {
-                            "id": (
-                                str(video.get("id"))
-                                if video.get("id") is not None
-                                else None
-                            ),
-                            "width": w,
-                            "height": h,
-                        },
-                    }
-                    video_items.append(item)
-                    break
-        return video_items
-    except Exception as e:
-        logger.error(
-            "pexels video search failed: "
-            f"error={type(e).__name__}, detail={_redact_request_error(e, api_key)}"
-        )
-
-    return []
-
-
-def search_videos_pixabay(
-    search_term: str,
-    minimum_duration: int,
-    video_aspect: VideoAspect = VideoAspect.portrait,
-) -> List[MaterialInfo]:
-    aspect = VideoAspect(video_aspect)
-
-    video_width, video_height = aspect.to_resolution()
-
-    api_key = get_api_key("pixabay_api_keys")
-    # Build URL
-    params = {
-        "q": search_term,
-        "video_type": "all",  # Accepted values: "all", "film", "animation"
-        "per_page": 50,
-        "key": api_key,
-    }
-    query_url = f"https://pixabay.com/api/videos/?{urlencode(params)}"
-    logger.info(
-        f"searching videos on pixabay: term={search_term!r}, "
-        f"proxy_enabled={bool(config.proxy)}"
-    )
-
-    try:
-        r = requests.get(
-            query_url, proxies=config.proxy, verify=_get_tls_verify(), timeout=(30, 60)
-        )
-        status_code = int(getattr(r, "status_code", 200))
-        headers = getattr(r, "headers", {}) or {}
-        content_type = str(headers.get("content-type", ""))
-        retry_after = headers.get("retry-after")
-        cf_ray = headers.get("cf-ray")
-
-        if _is_cloudflare_challenge(r):
-            logger.error(
-                "pixabay search was blocked by a Cloudflare challenge: "
-                f"status={status_code}, cf_ray={cf_ray or 'unknown'}. "
-                "Check the server network or proxy, or use Pexels/Coverr instead."
-            )
-            return []
-
-        if status_code == 429:
-            logger.error(
-                "pixabay API rate limit exceeded: "
-                f"status=429, retry_after={retry_after or 'unknown'}"
-            )
-            return []
-
-        if status_code >= 400:
-            logger.error(
-                "pixabay search request failed: "
-                f"status={status_code}, content_type={content_type or 'unknown'}"
-            )
-            return []
-
-        try:
-            response = r.json()
-        except ValueError:
-            logger.error(
-                "pixabay returned an unexpected non-JSON response: "
-                f"status={status_code}, content_type={content_type or 'unknown'}"
-            )
-            return []
-
-        video_items = []
-        if "hits" not in response:
-            logger.error("pixabay video search returned an unsupported response")
-            return video_items
-        videos = response["hits"]
-        # loop through each video in the result
-        for v in videos:
-            duration = v["duration"]
-            # check if video has desired minimum duration
-            if duration < minimum_duration:
-                continue
-            video_files = v["videos"]
-            # loop through each url to determine the best quality
-            for video_type in video_files:
-                video = video_files[video_type]
-                try:
-                    w = int(video["width"])
-                    h = int(video["height"])
-                except (KeyError, TypeError, ValueError):
-                    continue
-                # Pixabay 很少返回原生方形视频；1:1 输出继续接受满足分辨率的
-                # 候选并由合成阶段裁剪。横竖屏则必须严格匹配目标方向。
-                orientation_matches = aspect == VideoAspect.square or (
-                    _matches_video_aspect(w, h, aspect)
-                )
-                if orientation_matches and w >= video_width:
-                    item = MaterialInfo()
-                    item.provider = "pixabay"
-                    item.url = video["url"]
-                    item.duration = duration
-                    item.source_info = {
-                        "provider": "pixabay",
-                        "search_term": search_term,
-                        "asset_id": (
-                            str(v.get("id")) if v.get("id") is not None else None
-                        ),
-                        "source_page": _safe_public_url(v.get("pageURL")),
-                        "creator": _creator_info(
-                            {
-                                "id": v.get("user_id"),
-                                "name": v.get("user"),
-                            }
-                        ),
-                        "rendition": {
-                            "id": video_type,
-                            "width": w,
-                            "height": video.get("height"),
-                        },
-                    }
-                    video_items.append(item)
-                    break
-        return video_items
-    except Exception as e:
-        error_message = _redact_request_error(e, api_key)
-        logger.error(
-            "pixabay search request failed: "
-            f"error={type(e).__name__}, detail={error_message}"
-        )
-
-    return []
-
-
-def search_videos_coverr(
-    search_term: str,
-    minimum_duration: int,
-    video_aspect: VideoAspect = VideoAspect.portrait,
-) -> List[MaterialInfo]:
-    """
-    Coverr (https://coverr.co) - free HD/4K stock videos,
-    subject to Coverr license terms (https://coverr.co/license).
-
-    Coverr API notes (based on official docs at api.coverr.co/docs/):
-      - 鉴权: Authorization: Bearer <api_key>
-      - 搜索端点: GET /videos?query=...,响应结构 {"hits": [...], ...}
-      - 加 ?urls=true 在搜索响应里直接返回 mp4 直链
-      - URL 是 signed JWT(绑定 API key,无过期时间)
-      - Coverr 支持通过 filter=is_vertical:true/false 筛选横竖屏素材；
-        响应返回后仍根据 max_width/max_height 或 is_vertical 做本地校验
-      - duration 字段同时存在 number 和 string 两种形态,本函数都接受
-
-    本函数使用 urls.mp4_download 字段作为下载地址 —— 按 Coverr 官方文档
-    (https://api.coverr.co/docs/videos/#download-a-video) 的说法,
-    GET 这个 URL 本身就被 Coverr 当作一次合法的 download 事件计入统计,
-    无需再调用 PATCH /videos/:id/stats/downloads。
-    """
-    aspect = VideoAspect(video_aspect)
-    api_key = get_api_key("coverr_api_keys")
-    headers = {"Authorization": f"Bearer {api_key}"}
-    params = {
-        "query": search_term,
-        "page_size": 20,
-        "urls": "true",
-        "sort": "popular",
-    }
-    # 服务端方向筛选可以直接从完整搜索结果中返回目标素材，避免先取热门结果再
-    # 本地过滤导致竖屏候选为空。方形素材没有对应布尔条件，继续依赖本地宽高校验。
-    if aspect == VideoAspect.portrait:
-        params["filter"] = "is_vertical:true"
-    elif aspect == VideoAspect.landscape:
-        params["filter"] = "is_vertical:false"
-    query_url = f"https://api.coverr.co/videos?{urlencode(params)}"
-    logger.info(f"searching videos on coverr: term={search_term!r}")
-
-    try:
-        r = requests.get(
-            query_url,
-            headers=headers,
-            proxies=config.proxy,
-            verify=_get_tls_verify(),
-            timeout=(30, 60),
-        )
-        response = r.json()
-        video_items: List[MaterialInfo] = []
-
-        if not isinstance(response, dict) or "hits" not in response:
-            logger.error("coverr video search returned an unsupported response")
-            return video_items
-
-        for v in response["hits"]:
-            # duration 在不同响应里可能是 number(11.625) 或 string("10.500000")
-            try:
-                duration = int(float(v.get("duration") or 0))
-            except (TypeError, ValueError):
-                continue
-            if duration < minimum_duration:
-                continue
-
-            video_id = v.get("id")
-            mp4_download_url = (v.get("urls") or {}).get("mp4_download")
-            if not video_id or not mp4_download_url:
-                continue
-            if aspect != VideoAspect.square and not _matches_video_aspect(
-                v.get("max_width"),
-                v.get("max_height"),
-                aspect,
-                is_vertical=v.get("is_vertical"),
-            ):
-                continue
-
-            item = MaterialInfo()
-            item.provider = "coverr"
-            item.url = mp4_download_url
-            item.duration = duration
-            item.source_info = {
-                "provider": "coverr",
-                "search_term": search_term,
-                "asset_id": str(video_id),
-                "source_page": _safe_public_url(v.get("canonical_url") or v.get("url")),
-                "creator": _creator_info(v.get("creator") or v.get("author")),
-                "rendition": {
-                    "id": "mp4_download",
-                    "width": v.get("max_width"),
-                    "height": v.get("max_height"),
-                },
-            }
-            video_items.append(item)
-        return video_items
-    except Exception as e:
-        logger.error(
-            "coverr video search failed: "
-            f"error={type(e).__name__}, detail={_redact_request_error(e, api_key)}"
-        )
-
-    return []
-
-
-# WaveSpeed AI (https://wavespeed.ai) 通过文生视频模型按脚本关键词直接生成素材，
-# 与三个库存素材源共用 MaterialInfo 结果结构和后续下载、剪辑流程。
-WAVESPEED_API_BASE_URL = "https://api.wavespeed.ai/api/v3"
-WAVESPEED_DEFAULT_T2V_MODEL = "bytedance/seedance-2.0-fast/text-to-video"
-WAVESPEED_POLL_INTERVAL_SECONDS = 2.0
-WAVESPEED_RUN_TIMEOUT_SECONDS = 600.0
-# 默认模型 bytedance/seedance-2.0-fast/text-to-video 只接受 4-15 秒；超出
-# 范围的请求会被 API 直接拒绝。WebUI 默认片段时长是 3 秒，因此必须在提交
-# 前收敛到模型支持区间，多出的时长由现有剪辑流程按片段时长裁掉。
-WAVESPEED_MIN_DURATION_SECONDS = 4
-WAVESPEED_MAX_DURATION_SECONDS = 15
-# 三个失败态语义不同（模型报错 / 用户取消 / 平台超时），但对素材流程都意味着
-# 本关键词没有产物，统一按空结果处理，交给上层跳过该片段继续生成。
-WAVESPEED_FAILURE_STATUSES = frozenset({"failed", "cancelled", "timeout"})
-# 与 WaveSpeed 官方 Python SDK / n8n 节点保持同一口径：429 与 5xx 属于临时
-# 故障，值得有限次退避重试；4xx 是明确的客户端错误，快速失败。
-WAVESPEED_RETRYABLE_STATUS_CODES = frozenset({429, 500, 502, 503, 504})
-# 单次轮询允许的连续临时失败次数。一次不走运的 GET 不能让已经计费的任务失联。
-WAVESPEED_MAX_POLL_RETRIES = 5
-# 线性退避基数，第 n 次重试等待 base * n 秒。
-WAVESPEED_RETRY_BASE_SECONDS = 1.0
-# 产物下载失败时对同一个签名地址的重试次数。素材已经付费生成，优先重试原
-# 地址，不能因为一次下载抖动就重新提交一次付费生成任务。
-WAVESPEED_MAX_DOWNLOAD_RETRIES = 2
-
-
-class WaveSpeedUnconfirmedTaskError(RuntimeError):
-    """
-    付费生成任务已提交，但最终状态无法在本地确认。
-
-    这类异常绝不等价于“该任务失败、可以重来”：远端任务可能仍在运行或已经
-    完成并计费。素材流程必须就此停止，不再为后续关键词提交新的付费任务，
-    并把已提交的 prediction id 留在日志中供人工找回。
-    """
-
-    def __init__(self, message: str, prediction_id: str = ""):
-        super().__init__(message)
-        self.prediction_id = prediction_id
-
-
-def _wavespeed_status_code(response: Any) -> int:
-    """读取响应状态码；测试替身或异常对象缺少该字段时按 200 处理。"""
-    try:
-        return int(getattr(response, "status_code", 200))
-    except (TypeError, ValueError):
-        return 200
-
-
-def _is_wavespeed_retryable_error(error: Exception) -> bool:
-    """
-    判断轮询异常是否值得重试。
-
-    连接、超时一类网络异常没有状态码，按临时故障处理；带状态码的响应只在
-    429 和 5xx 时重试，与官方 SDK 的重试集合保持一致。
-    """
-    if isinstance(
-        error,
-        (
-            requests.exceptions.ConnectionError,
-            requests.exceptions.Timeout,
-            requests.exceptions.ChunkedEncodingError,
-        ),
-    ):
-        return True
-    response = getattr(error, "response", None)
-    if response is not None:
-        return _wavespeed_status_code(response) in WAVESPEED_RETRYABLE_STATUS_CODES
-    return False
-
-
-def _wavespeed_duration_bounds() -> tuple[int, int]:
-    """
-    返回当前模型支持的生成时长区间（秒）。
-
-    默认区间对应默认 Seedance 模型；用户切换到其它文生视频模型时，可以在
-    配置中同步调整区间。任何异常配置都退回默认值，并保证 min <= max，
-    避免把用户输入变成必然失败的远端请求。
-    """
-
-    def read_bound(key: str, fallback: int) -> int:
-        try:
-            value = int(config.app.get(key, fallback))
-        except (TypeError, ValueError):
-            return fallback
-        return value if value >= 1 else fallback
-
-    min_duration = read_bound("wavespeed_min_duration", WAVESPEED_MIN_DURATION_SECONDS)
-    max_duration = read_bound("wavespeed_max_duration", WAVESPEED_MAX_DURATION_SECONDS)
-    return min_duration, max(max_duration, min_duration)
-
-
-def generate_videos_wavespeed(
-    search_term: str,
-    minimum_duration: int,
-    video_aspect: VideoAspect = VideoAspect.portrait,
-) -> List[MaterialInfo]:
-    """
-    用 WaveSpeed 文生视频模型为一个脚本关键词生成一段素材。
-
-    与库存素材源的 search_videos_* 保持同一签名和空列表失败约定，
-    使其可以直接接入 ``download_videos`` 的通用下载与时长核算流程。
-    ``minimum_duration`` 在生成语境下就是目标片段时长（秒）。
-    """
-    aspect = VideoAspect(video_aspect)
-    video_width, video_height = aspect.to_resolution()
-    api_key = get_api_key("wavespeed_api_keys")
-    model_id = (
-        str(
-            config.app.get("wavespeed_text_to_video_model", "")
-            or WAVESPEED_DEFAULT_T2V_MODEL
-        )
-        .strip()
-        .strip("/")
-    )
-    headers = {"Authorization": f"Bearer {api_key}"}
-    requested_duration = max(int(minimum_duration), 1)
-    min_duration, max_duration = _wavespeed_duration_bounds()
-    duration = min(max(requested_duration, min_duration), max_duration)
-    if duration != requested_duration:
-        # 生成比请求更长不会影响成片：剪辑流程仍按片段时长裁剪；生成比请求
-        # 更短的情况只发生在请求超过模型上限时，此时也只能收敛到上限。
-        logger.info(
-            f"wavespeed clip duration clamped to model-supported range: "
-            f"requested={requested_duration}s, using={duration}s "
-            f"(supported {min_duration}-{max_duration}s)"
-        )
-    payload = {
-        "prompt": search_term,
-        "aspect_ratio": aspect.value,
-        "duration": duration,
-    }
-    logger.info(
-        f"generating video on wavespeed: model={model_id}, "
-        f"term={search_term!r}, duration={duration}s"
-    )
-
-    # 提交 POST 绝不自动重试：请求可能已经在远端创建了付费任务，重发会造成
-    # 重复生成和重复扣费（与官方 SDK 的 submission 策略一致）。
-    try:
-        submit_response = requests.post(
-            f"{WAVESPEED_API_BASE_URL}/{model_id}",
-            json=payload,
-            headers=headers,
-            proxies=config.proxy,
-            verify=_get_tls_verify(),
-            timeout=(30, 60),
-        )
-    except Exception as e:
-        # 没有收到响应并不代表任务没有创建。此时状态不明，必须终止整个生成
-        # 流程，而不是继续为下一个关键词提交新的付费任务。
-        raise WaveSpeedUnconfirmedTaskError(
-            "wavespeed submission did not return a response, the task may "
-            "already exist remotely: "
-            f"error={type(e).__name__}, detail={_redact_request_error(e, api_key)}"
-        ) from e
-
-    submit_status = _wavespeed_status_code(submit_response)
-    if submit_status >= 500:
-        # 5xx 可能发生在任务创建之后，无法判断是否已经计费。
-        raise WaveSpeedUnconfirmedTaskError(
-            f"wavespeed submission failed with HTTP {submit_status}, "
-            "the task may already exist remotely"
-        )
-    try:
-        submit_body = submit_response.json()
-    except Exception as e:
-        raise WaveSpeedUnconfirmedTaskError(
-            "wavespeed submission returned an unreadable response, the task "
-            f"may already exist remotely: error={type(e).__name__}"
-        ) from e
-
-    submit_data = submit_body.get("data") if isinstance(submit_body, dict) else None
-    if not isinstance(submit_body, dict) or submit_body.get("code") != 200:
-        # 4xx 与业务错误码是明确的拒绝，远端没有创建任务，也就不存在重复
-        # 计费风险，按现有素材源约定返回空结果并继续。
-        logger.error(
-            "wavespeed video generation request rejected: "
-            f"http_status={submit_status}, "
-            f"code={submit_body.get('code') if isinstance(submit_body, dict) else None}, "
-            f"detail={_redact_secret(str((submit_body or {}).get('message') or ''), api_key)}"
-        )
-        return []
-    prediction_id = (
-        str(submit_data.get("id") or "") if isinstance(submit_data, dict) else ""
-    )
-    if not prediction_id:
-        # 提交被接受但没拿到 ID：任务可能已经存在却无法追踪，不能继续下单。
-        raise WaveSpeedUnconfirmedTaskError(
-            "wavespeed accepted the submission without returning a prediction id"
-        )
-    # 生成任务提交成功即产生远端计费副作用，先落日志记录任务 ID，
-    # 即使后续轮询失败，用户仍能凭 ID 在 WaveSpeed 控制台找回产物。
-    logger.info(f"wavespeed prediction created: id={prediction_id}")
-
-    result_data = _wait_for_wavespeed_prediction(
-        prediction_id=prediction_id,
-        headers=headers,
-        api_key=api_key,
-    )
-    if result_data is None:
-        return []
-
-    try:
-        video_items = []
-        outputs = result_data.get("outputs")
-        for output in outputs if isinstance(outputs, list) else []:
-            # 产物 URL 是带签名的临时下载地址，必须整体保留（不能剥离查询参
-            # 数），因此不写入 source_info，只用于随后的立即下载。
-            if not isinstance(output, str) or not output.startswith(
-                ("http://", "https://")
-            ):
-                continue
-            item = MaterialInfo()
-            item.provider = "wavespeed"
-            item.url = output
-            item.duration = duration
-            item.source_info = {
-                "provider": "wavespeed",
-                "search_term": search_term,
-                "asset_id": prediction_id,
-                "rendition": {
-                    "id": None,
-                    "width": video_width,
-                    "height": video_height,
-                },
-            }
-            video_items.append(item)
-        if not video_items:
-            logger.error(
-                "wavespeed prediction completed without downloadable outputs: "
-                f"id={prediction_id}"
-            )
-        return video_items
-    except Exception as e:
-        # 产物已经生成并计费，这里的异常只可能来自本地解析。记录后按空结果
-        # 返回，让上层跳过该片段，但任务状态本身是确定的，可以继续后续片段。
-        logger.error(
-            "wavespeed output parsing failed: "
-            f"id={prediction_id}, error={type(e).__name__}, "
-            f"detail={_redact_request_error(e, api_key)}"
-        )
-
-    return []
-
-
-def _wait_for_wavespeed_prediction(
-    *,
-    prediction_id: str,
-    headers: dict,
-    api_key: str,
-) -> dict | None:
-    """
-    轮询同一个 prediction id 直到出现确定结果。
-
-    返回 ``completed`` 的 data；远端明确失败（failed / cancelled / timeout）
-    时返回 None，表示该任务已经结束、可以安全地继续后续片段。临时故障按
-    线性退避重试同一个 ID，绝不重新提交任务；状态始终无法确认时抛出
-    :class:`WaveSpeedUnconfirmedTaskError`，由调用方终止整个生成流程。
-    """
-    deadline = time.monotonic() + WAVESPEED_RUN_TIMEOUT_SECONDS
-    consecutive_failures = 0
-    while True:
-        try:
-            response = requests.get(
-                f"{WAVESPEED_API_BASE_URL}/predictions/{prediction_id}/result",
-                headers=headers,
-                proxies=config.proxy,
-                verify=_get_tls_verify(),
-                timeout=(30, 60),
-            )
-            status_code = _wavespeed_status_code(response)
-            if status_code in WAVESPEED_RETRYABLE_STATUS_CODES:
-                raise requests.exceptions.HTTPError(
-                    f"HTTP {status_code}", response=response
-                )
-            result_body = response.json()
-            result_data = (
-                result_body.get("data") if isinstance(result_body, dict) else None
-            )
-            if not isinstance(result_body, dict) or result_body.get("code") != 200:
-                # 轮询被明确拒绝（如 4xx）时任务状态仍然未知：任务已经提交，
-                # 只是本地查不到结果，同样不能继续提交新的付费任务。
-                raise WaveSpeedUnconfirmedTaskError(
-                    "wavespeed prediction status is unknown: "
-                    f"http_status={status_code}, "
-                    f"code={result_body.get('code') if isinstance(result_body, dict) else None}, "
-                    f"detail={_redact_secret(str((result_body or {}).get('message') or ''), api_key)}",
-                    prediction_id=prediction_id,
-                )
-            if not isinstance(result_data, dict):
-                raise WaveSpeedUnconfirmedTaskError(
-                    "wavespeed prediction result payload is malformed",
-                    prediction_id=prediction_id,
-                )
-        except WaveSpeedUnconfirmedTaskError:
-            raise
-        except Exception as e:
-            if not _is_wavespeed_retryable_error(e):
-                raise WaveSpeedUnconfirmedTaskError(
-                    "wavespeed prediction polling failed and the task state is "
-                    f"unknown: error={type(e).__name__}, "
-                    f"detail={_redact_request_error(e, api_key)}",
-                    prediction_id=prediction_id,
-                ) from e
-            consecutive_failures += 1
-            if consecutive_failures > WAVESPEED_MAX_POLL_RETRIES:
-                raise WaveSpeedUnconfirmedTaskError(
-                    "wavespeed prediction polling failed after "
-                    f"{WAVESPEED_MAX_POLL_RETRIES + 1} attempts, the task may "
-                    "still be running remotely: "
-                    f"error={type(e).__name__}, "
-                    f"detail={_redact_request_error(e, api_key)}",
-                    prediction_id=prediction_id,
-                ) from e
-            delay = WAVESPEED_RETRY_BASE_SECONDS * consecutive_failures
-            logger.warning(
-                "wavespeed prediction polling hit a transient error, retry the "
-                f"same task: id={prediction_id}, "
-                f"attempt={consecutive_failures}/{WAVESPEED_MAX_POLL_RETRIES}, "
-                f"error={type(e).__name__}, retry_in={delay:.1f}s"
-            )
-            time.sleep(delay)
-            continue
-
-        # 拿到一次有效响应就重置计数，只有连续失败才消耗重试额度。
-        consecutive_failures = 0
-        status = str(result_data.get("status") or "")
-        if status == "completed":
-            return result_data
-        if status in WAVESPEED_FAILURE_STATUSES:
-            logger.error(
-                "wavespeed prediction did not produce a video: "
-                f"id={prediction_id}, status={status}, "
-                f"detail={_redact_secret(str(result_data.get('error') or ''), api_key)}"
-            )
-            return None
-        if time.monotonic() > deadline:
-            # 远端任务仍在执行，本地无法确认最终状态，必须停止继续下单。
-            raise WaveSpeedUnconfirmedTaskError(
-                f"wavespeed prediction is still {status or 'pending'} after "
-                f"{WAVESPEED_RUN_TIMEOUT_SECONDS:.0f}s of local waiting",
-                prediction_id=prediction_id,
-            )
-        time.sleep(WAVESPEED_POLL_INTERVAL_SECONDS)
+# Provider implementations live in dedicated ``material_<provider>.py``
+# modules; this module stays the download/cache orchestration layer and a
+# thin facade so ``material.<name>`` keeps working for every caller (WebUI,
+# task pipeline, tests) exactly as before the split.
+from app.services.material_common import (  # noqa: F401
+    _creator_info,
+    _filter_materials_by_aspect,
+    _get_tls_verify,
+    _is_cloudflare_challenge,
+    _material_source_record,
+    _matches_video_aspect,
+    _persist_material_sources,
+    _redact_request_error,
+    _redact_secret,
+    _safe_public_url,
+    get_api_key,
+)
+from app.services.material_pexels import search_videos_pexels  # noqa: F401
+from app.services.material_pixabay import search_videos_pixabay  # noqa: F401
+from app.services.material_coverr import search_videos_coverr  # noqa: F401
+from app.services.material_wavespeed import (  # noqa: F401
+    WAVESPEED_MAX_DOWNLOAD_RETRIES,
+    WAVESPEED_MAX_POLL_RETRIES,
+    WAVESPEED_RETRY_BASE_SECONDS,
+    WAVESPEED_RUN_TIMEOUT_SECONDS,
+    WaveSpeedUnconfirmedTaskError,
+    generate_videos_wavespeed,
+)
 
 
 def _save_generated_video_with_retry(
