+++ app/services/material_common.py
@@ -1,0 +1,295 @@
+"""Shared helpers used by every material provider module under ``app.services.material_*``.
+
+Kept dependency-free from ``app.services.material`` and from any provider
+module, so the import graph stays a clean DAG: this module has no knowledge
+of who calls it, ``material.py`` and every ``material_<provider>.py`` import
+from here, and nothing imports back from those into this module.
+"""
+
+import threading
+from pathlib import Path
+from typing import Any
+from urllib.parse import quote_plus, urlsplit, urlunsplit
+
+import requests
+from loguru import logger
+
+from app.config import config
+from app.models.schema import MaterialInfo, VideoAspect
+from app.services import task_artifacts
+
+# Thread-safe counter for API key rotation
+_api_key_counter = 0
+_api_key_lock = threading.Lock()
+
+
+def _safe_public_url(value: Any) -> str | None:
+    """
+    只保留可公开展示的 HTTP(S) 页面地址，并移除查询参数和凭据。
+
+    素材下载地址可能携带 API Key、签名 JWT 或临时 token。任务清单只需要
+    帮助用户回到供应商的公开素材页，不应保存鉴权参数；用户信息形式的 URL
+    同样拒绝，避免 ``https://user:pass@example.com`` 一类内容落盘。
+    """
+    if not isinstance(value, str) or not value.strip():
+        return None
+
+    try:
+        parsed = urlsplit(value.strip())
+    except ValueError:
+        return None
+    if (
+        parsed.scheme not in {"http", "https"}
+        or not parsed.hostname
+        or parsed.username is not None
+        or parsed.password is not None
+    ):
+        return None
+    return urlunsplit((parsed.scheme, parsed.netloc, parsed.path, "", ""))
+
+
+def _creator_info(value: Any) -> dict[str, str] | None:
+    """从不同供应商的作者结构中提取统一的公开字段。"""
+    if isinstance(value, str) and value.strip():
+        return {"name": value.strip()}
+    if not isinstance(value, dict):
+        return None
+
+    creator: dict[str, str] = {}
+    creator_id = value.get("id")
+    creator_name = value.get("name") or value.get("username")
+    creator_page = _safe_public_url(
+        value.get("url") or value.get("profile_url") or value.get("profile_page")
+    )
+    if creator_id is not None:
+        creator["id"] = str(creator_id)
+    if creator_name:
+        creator["name"] = str(creator_name)
+    if creator_page:
+        creator["profile_page"] = creator_page
+    return creator or None
+
+
+def _material_source_record(item: MaterialInfo, local_path: str) -> dict[str, Any]:
+    """
+    为成功下载的素材生成轻量来源记录。
+
+    ``source_info`` 可能来自缓存，甚至来自外部构造的 ``MaterialInfo``，因此
+    不能原样写入。这里按白名单重新构造，只保留公开页面、业务标识和尺寸，
+    并只记录本地文件名，避免用户目录或 Docker 挂载路径进入任务文件。
+    """
+    source = item.source_info if isinstance(item.source_info, dict) else {}
+    record: dict[str, Any] = {
+        "provider": str(item.provider or source.get("provider") or ""),
+        "local_file": Path(local_path).name,
+        "duration": int(item.duration),
+    }
+
+    search_term = source.get("search_term")
+    asset_id = source.get("asset_id")
+    source_page = _safe_public_url(source.get("source_page"))
+    if isinstance(search_term, str) and search_term.strip():
+        record["search_term"] = search_term.strip()
+    if asset_id not in (None, ""):
+        record["asset_id"] = str(asset_id)
+    if source_page:
+        record["source_page"] = source_page
+
+    creator = _creator_info(source.get("creator"))
+    if creator:
+        record["creator"] = creator
+
+    raw_rendition = source.get("rendition")
+    if isinstance(raw_rendition, dict):
+        rendition = {}
+        for field in ("id", "width", "height"):
+            value = raw_rendition.get(field)
+            if value not in (None, ""):
+                rendition[field] = str(value) if field == "id" else value
+        if rendition:
+            record["rendition"] = rendition
+    return record
+
+
+def _persist_material_sources(
+    task_id: str,
+    material_sources: list[dict[str, Any]],
+) -> None:
+    """
+    将当前实际下载成功的素材来源补充到任务清单。
+
+    任务记录是辅助能力，不能改变视频下载函数的返回值，也不能因为写盘失败
+    中断成片主流程。``patch_script_data`` 会负责原子替换和异常日志；这里仅在
+    成功后记录数量，便于确认任务追溯信息是否已经落盘。
+    """
+    try:
+        saved = task_artifacts.patch_script_data(
+            task_id,
+            material_sources=material_sources,
+        )
+        if saved:
+            logger.info(
+                f"saved material source records: "
+                f"task_id={task_id}, count={len(material_sources)}"
+            )
+    except Exception as exc:
+        # task_artifacts 自身已经按失败降级设计，这里仍保留最后一道隔离，
+        # 防止未来实现调整或目录解析异常意外影响素材下载返回值。
+        logger.warning(
+            "failed to persist material source records: "
+            f"task_id={task_id}, error={type(exc).__name__}, detail={exc}"
+        )
+
+
+def _get_tls_verify() -> bool:
+    # 默认开启 TLS 证书校验，防止素材搜索和下载过程被中间人篡改。
+    # 仅在企业代理、自签证书等明确需要的场景下，允许用户通过
+    # `config.toml` 显式设置 `tls_verify = false` 临时关闭。
+    tls_verify = config.app.get("tls_verify", True)
+    if isinstance(tls_verify, str):
+        tls_verify = tls_verify.strip().lower() not in ("0", "false", "no", "off")
+
+    if not tls_verify:
+        logger.warning(
+            "TLS certificate verification is disabled by config.app.tls_verify=false. "
+            "Only use this in trusted proxy environments."
+        )
+
+    return bool(tls_verify)
+
+
+def get_api_key(cfg_key: str):
+    api_keys = config.app.get(cfg_key)
+    if not api_keys:
+        raise ValueError(
+            f"\n\n##### {cfg_key} is not set #####\n\n"
+            f"Please set it in the config.toml file: {config.config_file}\n"
+        )
+
+    # if only one key is provided, return it
+    if isinstance(api_keys, str):
+        return api_keys
+
+    global _api_key_counter
+    with _api_key_lock:
+        _api_key_counter += 1
+        return api_keys[_api_key_counter % len(api_keys)]
+
+
+def _redact_secret(message: str, secret: str) -> str:
+    """
+    对即将写入日志的异常文本做最小范围脱敏。
+
+    requests 的连接异常可能包含完整请求 URL，而 Pixabay API Key 通过查询
+    参数传递。这里同时替换原始值和 URL 编码值，既保留网络错误信息用于排查，
+    又避免密钥进入日志文件。
+    """
+    safe_message = str(message)
+    if not secret:
+        return safe_message
+
+    safe_message = safe_message.replace(secret, "***")
+    encoded_secret = quote_plus(secret)
+    if encoded_secret != secret:
+        safe_message = safe_message.replace(encoded_secret, "***")
+    return safe_message
+
+
+def _redact_request_error(error: Exception, *secrets: str) -> str:
+    """
+    保留网络异常的可排查信息，同时移除 API Key 和代理凭据。
+
+    直接只记录异常类型会丢失 DNS、证书、超时等关键上下文；直接记录原始异常
+    又可能回显完整请求 URL。统一入口可以让三个素材供应商使用相同脱敏规则。
+    """
+    safe_message = str(error)
+    for secret in secrets:
+        safe_message = _redact_secret(safe_message, str(secret or ""))
+    for proxy_url in config.proxy.values():
+        safe_message = _redact_secret(safe_message, str(proxy_url))
+    return safe_message
+
+
+def _is_cloudflare_challenge(response: requests.Response) -> bool:
+    """
+    识别 Cloudflare 返回的 HTML Challenge，而不是把它当成 Pixabay JSON。
+
+    Cloudflare 通常会设置 `cf-mitigated: challenge`；部分部署只返回带有
+    "Just a moment" 或 challenge-platform 的 HTML，因此保留内容特征兜底。
+    响应正文仅在内存中判断，不写入日志，避免记录无价值的大段 HTML。
+    """
+    headers = getattr(response, "headers", {}) or {}
+    if str(headers.get("cf-mitigated", "")).lower() == "challenge":
+        return True
+
+    content_type = str(headers.get("content-type", "")).lower()
+    if "text/html" not in content_type:
+        return False
+
+    body = str(getattr(response, "text", "")).lower()
+    return "just a moment" in body or "/cdn-cgi/challenge-platform/" in body
+
+
+def _matches_video_aspect(
+    width: Any,
+    height: Any,
+    video_aspect: VideoAspect,
+    *,
+    is_vertical: Any = None,
+) -> bool:
+    """
+    判断远端素材是否与目标画面方向一致。
+
+    Pexels、Pixabay 和 Coverr 的响应字段并不统一，因此先使用宽高做可靠判断；
+    Coverr 部分历史响应缺少尺寸时，再使用明确的 ``is_vertical`` 布尔值兜底。
+    无法确认方向的素材直接跳过，避免竖屏任务混入横屏素材并在成片中产生黑边。
+    """
+    aspect = VideoAspect(video_aspect)
+    try:
+        normalized_width = int(float(width))
+        normalized_height = int(float(height))
+    except (TypeError, ValueError):
+        normalized_width = 0
+        normalized_height = 0
+
+    if normalized_width > 0 and normalized_height > 0:
+        if aspect == VideoAspect.portrait:
+            return normalized_height > normalized_width
+        if aspect == VideoAspect.landscape:
+            return normalized_width > normalized_height
+        return normalized_width == normalized_height
+
+    if isinstance(is_vertical, bool) and aspect != VideoAspect.square:
+        return is_vertical == (aspect == VideoAspect.portrait)
+    return False
+
+
+def _filter_materials_by_aspect(
+    items: list[MaterialInfo],
+    video_aspect: VideoAspect,
+) -> list[MaterialInfo]:
+    """
+    对缓存结果再次校验方向。
+
+    素材搜索缓存最长保留 24 小时，升级前写入的缓存可能包含方向不匹配的素材。
+    在统一缓存入口过滤可以让修复立即生效，也能防御第三方 Provider 或旧缓存
+    遗漏远端筛选。无法读取 rendition 尺寸的旧条目按未验证处理并跳过。
+    """
+    aspect = VideoAspect(video_aspect)
+    if aspect == VideoAspect.square:
+        # Pixabay 和 Coverr 很少提供原生方形素材。方形输出沿用既有行为，
+        # 接受可用候选并交给视频合成阶段裁剪，避免升级后 1:1 任务无素材。
+        return list(items)
+
+    filtered_items = []
+    for item in items:
+        source_info = item.source_info if isinstance(item.source_info, dict) else {}
+        rendition = source_info.get("rendition")
+        rendition = rendition if isinstance(rendition, dict) else {}
+        if _matches_video_aspect(
+            rendition.get("width"),
+            rendition.get("height"),
+            aspect,
+        ):
+            filtered_items.append(item)
+    return filtered_items
