+++ app/services/material_coverr.py
@@ -1,0 +1,122 @@
+"""Coverr stock video search provider."""
+
+from typing import List
+from urllib.parse import urlencode
+
+import requests
+from loguru import logger
+
+from app.config import config
+from app.models.schema import MaterialInfo, VideoAspect
+from app.services.material_common import (
+    _creator_info,
+    _get_tls_verify,
+    _matches_video_aspect,
+    _redact_request_error,
+    _safe_public_url,
+    get_api_key,
+)
+
+
+def search_videos_coverr(
+    search_term: str,
+    minimum_duration: int,
+    video_aspect: VideoAspect = VideoAspect.portrait,
+) -> List[MaterialInfo]:
+    """
+    Coverr (https://coverr.co) - free HD/4K stock videos,
+    subject to Coverr license terms (https://coverr.co/license).
+
+    Coverr API notes (based on official docs at api.coverr.co/docs/):
+      - 鉴权: Authorization: Bearer <api_key>
+      - 搜索端点: GET /videos?query=...,响应结构 {"hits": [...], ...}
+      - 加 ?urls=true 在搜索响应里直接返回 mp4 直链
+      - URL 是 signed JWT(绑定 API key,无过期时间)
+      - Coverr 支持通过 filter=is_vertical:true/false 筛选横竖屏素材；
+        响应返回后仍根据 max_width/max_height 或 is_vertical 做本地校验
+      - duration 字段同时存在 number 和 string 两种形态,本函数都接受
+
+    本函数使用 urls.mp4_download 字段作为下载地址 —— 按 Coverr 官方文档
+    (https://api.coverr.co/docs/videos/#download-a-video) 的说法,
+    GET 这个 URL 本身就被 Coverr 当作一次合法的 download 事件计入统计,
+    无需再调用 PATCH /videos/:id/stats/downloads。
+    """
+    aspect = VideoAspect(video_aspect)
+    api_key = get_api_key("coverr_api_keys")
+    headers = {"Authorization": f"Bearer {api_key}"}
+    params = {
+        "query": search_term,
+        "page_size": 20,
+        "urls": "true",
+        "sort": "popular",
+    }
+    # 服务端方向筛选可以直接从完整搜索结果中返回目标素材，避免先取热门结果再
+    # 本地过滤导致竖屏候选为空。方形素材没有对应布尔条件，继续依赖本地宽高校验。
+    if aspect == VideoAspect.portrait:
+        params["filter"] = "is_vertical:true"
+    elif aspect == VideoAspect.landscape:
+        params["filter"] = "is_vertical:false"
+    query_url = f"https://api.coverr.co/videos?{urlencode(params)}"
+    logger.info(f"searching videos on coverr: term={search_term!r}")
+
+    try:
+        r = requests.get(
+            query_url,
+            headers=headers,
+            proxies=config.proxy,
+            verify=_get_tls_verify(),
+            timeout=(30, 60),
+        )
+        response = r.json()
+        video_items: List[MaterialInfo] = []
+
+        if not isinstance(response, dict) or "hits" not in response:
+            logger.error("coverr video search returned an unsupported response")
+            return video_items
+
+        for v in response["hits"]:
+            # duration 在不同响应里可能是 number(11.625) 或 string("10.500000")
+            try:
+                duration = int(float(v.get("duration") or 0))
+            except (TypeError, ValueError):
+                continue
+            if duration < minimum_duration:
+                continue
+
+            video_id = v.get("id")
+            mp4_download_url = (v.get("urls") or {}).get("mp4_download")
+            if not video_id or not mp4_download_url:
+                continue
+            if aspect != VideoAspect.square and not _matches_video_aspect(
+                v.get("max_width"),
+                v.get("max_height"),
+                aspect,
+                is_vertical=v.get("is_vertical"),
+            ):
+                continue
+
+            item = MaterialInfo()
+            item.provider = "coverr"
+            item.url = mp4_download_url
+            item.duration = duration
+            item.source_info = {
+                "provider": "coverr",
+                "search_term": search_term,
+                "asset_id": str(video_id),
+                "source_page": _safe_public_url(v.get("canonical_url") or v.get("url")),
+                "creator": _creator_info(v.get("creator") or v.get("author")),
+                "rendition": {
+                    "id": "mp4_download",
+                    "width": v.get("max_width"),
+                    "height": v.get("max_height"),
+                },
+            }
+            video_items.append(item)
+        return video_items
+    except Exception as e:
+        logger.error(
+            "coverr video search failed: "
+            f"error={type(e).__name__}, detail={_redact_request_error(e, api_key)}"
+        )
+
+    return []
