+++ app/services/material_pexels.py
@@ -1,0 +1,100 @@
+"""Pexels stock video search provider."""
+
+from typing import List
+from urllib.parse import urlencode
+
+import requests
+from loguru import logger
+
+from app.config import config
+from app.models.schema import MaterialInfo, VideoAspect
+from app.services.material_common import (
+    _creator_info,
+    _get_tls_verify,
+    _matches_video_aspect,
+    _redact_request_error,
+    _safe_public_url,
+    get_api_key,
+)
+
+
+def search_videos_pexels(
+    search_term: str,
+    minimum_duration: int,
+    video_aspect: VideoAspect = VideoAspect.portrait,
+) -> List[MaterialInfo]:
+    aspect = VideoAspect(video_aspect)
+    video_orientation = aspect.name
+    video_width, video_height = aspect.to_resolution()
+    api_key = get_api_key("pexels_api_keys")
+    headers = {
+        "Authorization": api_key,
+        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/115.0.0.0 Safari/537.36",
+    }
+    # Build URL
+    params = {"query": search_term, "per_page": 20, "orientation": video_orientation}
+    query_url = f"https://api.pexels.com/v1/videos/search?{urlencode(params)}"
+    logger.info(f"searching videos on pexels: term={search_term!r}")
+
+    try:
+        r = requests.get(
+            query_url,
+            headers=headers,
+            proxies=config.proxy,
+            verify=_get_tls_verify(),
+            timeout=(30, 60),
+        )
+        response = r.json()
+        video_items = []
+        if "videos" not in response:
+            logger.error("pexels video search returned an unsupported response")
+            return video_items
+        videos = response["videos"]
+        # loop through each video in the result
+        for v in videos:
+            duration = v["duration"]
+            # check if video has desired minimum duration
+            if duration < minimum_duration:
+                continue
+            video_files = v["video_files"]
+            # loop through each url to determine the best quality
+            for video in video_files:
+                w = int(video["width"])
+                h = int(video["height"])
+                if (
+                    _matches_video_aspect(w, h, aspect)
+                    and w == video_width
+                    and h == video_height
+                ):
+                    item = MaterialInfo()
+                    item.provider = "pexels"
+                    item.url = video["link"]
+                    item.duration = duration
+                    item.source_info = {
+                        "provider": "pexels",
+                        "search_term": search_term,
+                        "asset_id": (
+                            str(v.get("id")) if v.get("id") is not None else None
+                        ),
+                        "source_page": _safe_public_url(v.get("url")),
+                        "creator": _creator_info(v.get("user")),
+                        "rendition": {
+                            "id": (
+                                str(video.get("id"))
+                                if video.get("id") is not None
+                                else None
+                            ),
+                            "width": w,
+                            "height": h,
+                        },
+                    }
+                    video_items.append(item)
+                    break
+        return video_items
+    except Exception as e:
+        logger.error(
+            "pexels video search failed: "
+            f"error={type(e).__name__}, detail={_redact_request_error(e, api_key)}"
+        )
+
+    return []
