+++ app/services/material_pixabay.py
@@ -1,0 +1,145 @@
+"""Pixabay stock video search provider."""
+
+from typing import List
+from urllib.parse import urlencode
+
+import requests
+from loguru import logger
+
+from app.config import config
+from app.models.schema import MaterialInfo, VideoAspect
+from app.services.material_common import (
+    _creator_info,
+    _get_tls_verify,
+    _is_cloudflare_challenge,
+    _matches_video_aspect,
+    _redact_request_error,
+    _safe_public_url,
+    get_api_key,
+)
+
+
+def search_videos_pixabay(
+    search_term: str,
+    minimum_duration: int,
+    video_aspect: VideoAspect = VideoAspect.portrait,
+) -> List[MaterialInfo]:
+    aspect = VideoAspect(video_aspect)
+
+    video_width, video_height = aspect.to_resolution()
+
+    api_key = get_api_key("pixabay_api_keys")
+    # Build URL
+    params = {
+        "q": search_term,
+        "video_type": "all",  # Accepted values: "all", "film", "animation"
+        "per_page": 50,
+        "key": api_key,
+    }
+    query_url = f"https://pixabay.com/api/videos/?{urlencode(params)}"
+    logger.info(
+        f"searching videos on pixabay: term={search_term!r}, "
+        f"proxy_enabled={bool(config.proxy)}"
+    )
+
+    try:
+        r = requests.get(
+            query_url, proxies=config.proxy, verify=_get_tls_verify(), timeout=(30, 60)
+        )
+        status_code = int(getattr(r, "status_code", 200))
+        headers = getattr(r, "headers", {}) or {}
+        content_type = str(headers.get("content-type", ""))
+        retry_after = headers.get("retry-after")
+        cf_ray = headers.get("cf-ray")
+
+        if _is_cloudflare_challenge(r):
+            logger.error(
+                "pixabay search was blocked by a Cloudflare challenge: "
+                f"status={status_code}, cf_ray={cf_ray or 'unknown'}. "
+                "Check the server network or proxy, or use Pexels/Coverr instead."
+            )
+            return []
+
+        if status_code == 429:
+            logger.error(
+                "pixabay API rate limit exceeded: "
+                f"status=429, retry_after={retry_after or 'unknown'}"
+            )
+            return []
+
+        if status_code >= 400:
+            logger.error(
+                "pixabay search request failed: "
+                f"status={status_code}, content_type={content_type or 'unknown'}"
+            )
+            return []
+
+        try:
+            response = r.json()
+        except ValueError:
+            logger.error(
+                "pixabay returned an unexpected non-JSON response: "
+                f"status={status_code}, content_type={content_type or 'unknown'}"
+            )
+            return []
+
+        video_items = []
+        if "hits" not in response:
+            logger.error("pixabay video search returned an unsupported response")
+            return video_items
+        videos = response["hits"]
+        # loop through each video in the result
+        for v in videos:
+            duration = v["duration"]
+            # check if video has desired minimum duration
+            if duration < minimum_duration:
+                continue
+            video_files = v["videos"]
+            # loop through each url to determine the best quality
+            for video_type in video_files:
+                video = video_files[video_type]
+                try:
+                    w = int(video["width"])
+                    h = int(video["height"])
+                except (KeyError, TypeError, ValueError):
+                    continue
+                # Pixabay 很少返回原生方形视频；1:1 输出继续接受满足分辨率的
+                # 候选并由合成阶段裁剪。横竖屏则必须严格匹配目标方向。
+                orientation_matches = aspect == VideoAspect.square or (
+                    _matches_video_aspect(w, h, aspect)
+                )
+                if orientation_matches and w >= video_width:
+                    item = MaterialInfo()
+                    item.provider = "pixabay"
+                    item.url = video["url"]
+                    item.duration = duration
+                    item.source_info = {
+                        "provider": "pixabay",
+                        "search_term": search_term,
+                        "asset_id": (
+                            str(v.get("id")) if v.get("id") is not None else None
+                        ),
+                        "source_page": _safe_public_url(v.get("pageURL")),
+                        "creator": _creator_info(
+                            {
+                                "id": v.get("user_id"),
+                                "name": v.get("user"),
+                            }
+                        ),
+                        "rendition": {
+                            "id": video_type,
+                            "width": w,
+                            "height": video.get("height"),
+                        },
+                    }
+                    video_items.append(item)
+                    break
+        return video_items
+    except Exception as e:
+        error_message = _redact_request_error(e, api_key)
+        logger.error(
+            "pixabay search request failed: "
+            f"error={type(e).__name__}, detail={error_message}"
+        )
+
+    return []
